requirements.txt
asgiref==3.3.1
boto3==1.17.49
botocore==1.20.49
dj-database-url==0.5.0
Django==3.1.7
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
